refactor(dashboard): remove some props from gameLibraryGrid
